Ready for restructure; compiling once more.
This is not doing what it needs to do _yet_, but it's ready for
the rework. I've erased the entire `md2notion` crate and moved all
of its `convert()` function into the migrator tool. The convert
function now requires a client, so that it can split up hierarchies
that go too deep when it detects them, and migrate deep page
structures in chunks.

The convert function is _not yet_ doing this, but the bottleneck
is there, ready and waiting for it.